requirements.txt
google-generativeai
qdrant-client
cohere
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-community
